Removed with_context from various file names. Merged 1.23 result tests with result tests.
